pyproject.toml
[tool.poetry]
name = "arest"
version = "0.1.0"
description = ""
authors = ["ilia <kavaleuilia@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = 'arest', from = 'src/'},
    { include = 'log', from = 'src/'},
]

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.9.1"
pyyaml = "^6.0.1"
asyncpg = "^0.29.0"
aiohttp-jinja2 = "^1.6"
dynaconf = "^3.2.4"
controller = {git = "https://gitlab.com/xolloo/aiohttp_controller.git"}
aiohttp-cors = "^0.7.0"
tortoise-orm = "^0.20.0"
aerich = "^0.7.2"
pyjwt = "^2.8.0"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
aiohttp-devtools = "^1.1.2"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"



[tool.aerich]
tortoise_orm = "src.arest.conf._conf.DATABASE"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
